requirements.txt
requests==2.31.0
moviepy==1.0.3
pillow==10.1.0
instabot==0.117.0
python-dotenv==1.0.0
gTTS==2.4.0
arabic-reshaper==3.0.0
python-bidi==0.4.2
ffmpeg-python==0.2.0
pydub==0.25.1
